Add BBSComponent
